Moving CW Insights tabs to CW Logs toolwindow
