+++ jetbrains-core/resources/META-INF/plugin.xml
@@ -175,7 +175,6 @@ with what features/services are supported.
         <projectService serviceImplementation="software.aws.toolkits.jetbrains.core.toolwindow.ToolkitToolWindowManager" />
         <projectService serviceImplementation="software.aws.toolkits.jetbrains.services.cloudwatch.logs.CloudWatchLogWindow" />
         <projectService serviceImplementation="software.aws.toolkits.jetbrains.services.sqs.toolwindow.SqsWindow" />
-        <projectService serviceImplementation="software.aws.toolkits.jetbrains.services.cloudwatch.logs.insights.QueryResultsWindow" />
       
         <toolWindow id="aws.explorer" anchor="left" secondary="true"
                     factoryClass="software.aws.toolkits.jetbrains.core.explorer.AwsExplorerFactory"
