+++ resources/resources/software/aws/toolkits/resources/localized_messages.properties
@@ -247,7 +247,7 @@ cloudwatch.logs.query_result_completion_status=Result
 cloudwatch.logs.query_result_completion_successful=Query Execution Completed
 cloudwatch.logs.query_results_table_error=Failed to load Query Results
 cloudwatch.logs.query_saved_successfully=Query has been saved to your account!
-cloudwatch.logs.results_window_title=CloudWatch Logs Insights
+cloudwatch.logs.query_tab_title=Query: {0}
 cloudwatch.logs.save_action=Save to a File
 cloudwatch.logs.save_query=Save Query
 cloudwatch.logs.save_query_dialog_name=Enter Query Name
